Add LaunchMostRecentAppImage functionality and "update" command
LaunchMostRecentAppImage launches an the most recent application for a given
updateinformation that we found among the integrated AppImages.
Kinda like poor man's Launch Services.

Probably we should make as much use of it as possible.

I think this way of doing things is really clever because
this way we can even put the update action into menus if
no updater is on the system yet at that time, and give the user
instructions what to do once they want to use the upate action
in the menu. Also, this way the updater does not have to be at
a static location on the $PATH but can be put into any location
from which it gets integrated.