Fix arch detection (#81)
* fix: detection of ARCH from libraries

Previously, go-appimagetool did not stop if a proper architecture was
not detected from the .so. files or the binary. As it did not halt,
the arch variable was set to empty string (''). This led to a
filenotfound panic, because the tool was looking for
`release-` instead of `release-x86_64` as the arch variable was
empty. go-appimagetoold should properly halt if the detection of
ARCH variable did not complete successfully, so that they
can provide the ARCH environment variable instead.

* style: run go fmt