Make servicemenu available with newer KDE Frameworks (#354)
Since KDE Frameworks 5.85, servicemenus are located at
`$XDG_DATA_HOME/kio/servicemenus`.
The appimaged servicemenu desktop file is kept in the original location, and
symlinks are added in new locations. This way, users of older versions of KDE
apps are still supported, and newer servicemenu locations can be easily
supported (if it is changed again in the future).