Update systemd service unit file
Update systemd unit file for appimaged service

- Add `Slice=background.slice` to allow systemd to deprioritize appimaged
background service, as AppImage daemon does not provide critical or
time-sensitive services
- New line at the end of appimaged service unit file
- Fix race condition between AppImage daemon and GVfs/KDE Solid UDisks2 volume
monitor

----------------------------------------

AppImage daemon wants to ensure that signals from
`org.gtk.vfs.UDisks2VolumeMonitor` and `org.kde.Solid.Device` would be sent, so
it checks for relevant names on the bus.
For the names to definitely exist on the bus when AppImage daemon checks, it
needs to start later than the services acquiring the names.

Both `gvfs-udisks2-volume-monitor.service` and `plasma-powerdevil.service`
are "static" services (No `[Install]` section, only started if other units
depend on them).
Taking PowerDevil as example, it is depended on by `plasma-workspace.target`.
`plasma-workspace-{x11/wayland}.target` are in turn also static, and only
requested to be started by `startplasma-{x11/wayland}` binaries, which is
executed by SDDM (the display manager) when the user chooses a session and logs
into it.

-----

Previously, appimaged service unit uses `WantedBy=default.target`.
**`default.target` of user units is started when the user service manager is
started.**
(Not to be confused with `default.target` of system units, which is started on
boot, and is often aliased to `multi-user.target` or `graphical.target` system
units.)
The user service manager is normally started at user login. If "user lingering"
is enabled, the user service manager is started as soon as possible after boot.

-----

From the above, we can understand that:

1. Without user lingering:

- System boots, system service manager is started.
- At the login screen, the user chooses a session and enters the password to
log in.
- The user service manager is started, and starts `default.target` and its
dependencies - including `appimaged.service`.
- The binary set as the value of the session's `Exec=` key (for example
`startplasma-wayland`) is executed, and is responsible for setting up and
starting the rest of the session. It requests the proper user service to be
started (`plasma-workspace-wayland.target`).
- The user service manager starts that session-specific service
(`plasma-workspace-wayland.target`) and its dependencies - including
`plasma-powerdevil.service`.

At this point, both AppImage daemon and PowerDevil are starting up.

1.a. If AppImage daemon runs `monitorUdisks()` **before** PowerDevil acquires
the name on the bus, AppImage daemon assumes that volumes can not be monitored,
and sends a notification "Cannot see volumes come and go".
1.b. If AppImage daemon runs `monitorUdisks()` **after** PowerDevil acquires
the name on the bus, everything is fine.

2. With user lingering:

- System boots, system service manager is started.
- The user service manager is started, and starts `default.target` and its
dependencies - including `appimaged.service`.
- AppImage daemon starts up, runs `monitorUdisks()` before PowerDevil is even
started, assumes that volumes can not be monitored, and sends a notification
"Cannot see volumes come and go".
- At the login screen, the user chooses a session and enters the password to
log in.
- The binary set as the value of the session's `Exec=` key (for example
`startplasma-wayland`) is executed, and is responsible for setting up and
starting the rest of the session. It requests the proper user service to be
started (`plasma-workspace-wayland.target`).
- The user service manager starts that session-specific service
(`plasma-workspace-wayland.target`) and its dependencies - including
`plasma-powerdevil.service`.

-----

Without user lingering, a race condition exists between appimaged service and
volume monitoring service.
With user lingering, appimaged service is always started before volume
monitoring service.

To solve the problem, `gvfs-udisks2-volume-monitor.service` and
`plasma-powerdevil.service` are added to `WantedBy=` *in addition to*
`default.target`.
Instead of starting AppImage daemon and checking/waiting for the availability
of volume monitoring services, now AppImage daemon is started *because of*
those services.
As well, they're added to `After=` so that appimaged service is definitely
started after them if they succeed.
`ExecCondition=` is added to skip appimaged service being started, when volume
monitoring services aren't ready yet, but are reasonably predicted to have
started at a near future point in time.

The new steps are:

1. System boots, system service manager is started.
2. Either now or after login, the user service manager is started, and starts
`default.target` and its dependencies - including `appimaged.service`.
3.a. If at least one of `gvfs-udisks2-volume-monitor.service` and
`plasma-powerdevil.service` exists and is enabled, but none are active, skip
starting `appimaged.service` now.
3.b. Else, AppImage daemon starts up, runs `monitorUdisks()`, and either
3.b.(a) enabled && active: succeeds, or
3.b.(b) !enabled: rightfully determines that volumes can not be monitored, at
least in the current implementation of AppImage daemon, and sends the
notification "Cannot see volumes come and go".
4. At the login screen, the user chooses a session and enters the password to
log in.
5. The binary set as the value of the session's `Exec=` key (for example
`startplasma-wayland`) is executed, and is responsible for setting up and
starting the rest of the session. It requests the proper user service to be
started (`plasma-workspace-wayland.target`).
6. The user service manager starts that session-specific service
(`plasma-workspace-wayland.target`) and its dependencies - including
`plasma-powerdevil.service`.
7. The volume monitoring service (`plasma-powerdevil.service`) is set to want
and start before `appimaged.service`, and either starts successfully or fails.
8. AppImage daemon starts up, runs `monitorUdisks()`, and either succeeds, or
rightfully determines that volumes can not be monitored, at least in the
current implementation of AppImage daemon, and sends the notification "Cannot
see volumes come and go".