Fix appimaged service unit file properly
See #355 for details

Found the root cause of unexpected race condition with PowerDevil service,
removed unnecessary `WantedBy=` dependency and `After=` ordering on
`gvfs-udisks2-volume-monitor.service` as its D-Bus activated.

Code related to UDisks should be reworked, as PowerDevil is unrelated to
volume monitoring. However for now this is probably the best solution to the
erroneous detection and "Cannot see volumes come and go" notification issue.